[WIP] P-Chain merkle DB -- smaller diff (#2382)
