+++ vms/platformvm/state/masked_iterator_test.go
@@ -17,23 +17,19 @@ func TestMaskedIterator(t *testing.T) {
 	stakers := []*Staker{
 		{
 			TxID:     ids.GenerateTestID(),
-			Weight:   0, // just to simplify debugging
 			NextTime: time.Unix(0, 0),
 		},
 		{
 			TxID:     ids.GenerateTestID(),
-			Weight:   10, // just to simplify debugging
-			NextTime: time.Unix(10, 0),
+			NextTime: time.Unix(1, 0),
 		},
 		{
 			TxID:     ids.GenerateTestID(),
-			Weight:   20, // just to simplify debugging
-			NextTime: time.Unix(20, 0),
+			NextTime: time.Unix(2, 0),
 		},
 		{
 			TxID:     ids.GenerateTestID(),
-			Weight:   30, // just to simplify debugging
-			NextTime: time.Unix(30, 0),
+			NextTime: time.Unix(3, 0),
 		},
 	}
 	maskedStakers := map[ids.ID]*Staker{
