+++ vms/platformvm/vm_regression_test.go
@@ -1436,6 +1436,7 @@ func TestSubnetValidatorBLSKeyDiffAfterExpiry(t *testing.T) {
 	vm.ctx.Lock.Lock()
 	defer func() {
 		require.NoError(vm.Shutdown(context.Background()))
+
 		vm.ctx.Lock.Unlock()
 	}()
 	subnetID := testSubnet1.TxID
